Componentize layout more
